+++ testdata/panic_const_division/input.go
@@ -1,0 +1,4 @@
+package constdivision
+
+//gents:export
+const Third = 10 / 3