+++ testdata/panic_const_forward_ref/input.go
@@ -1,0 +1,8 @@
+package constforwardref
+
+//gents:export
+const (
+	All = A | B
+	A   = 1
+	B   = 2
+)