+++ testdata/panic_const_iota/input.go
@@ -1,0 +1,7 @@
+package constiota
+
+//gents:export
+const (
+	A = iota
+	B
+)