+++ testdata/panic_marker_near_miss/input.go
@@ -1,0 +1,6 @@
+package markernearmiss
+
+// gents:export
+type Foo struct {
+	A string `json:"a"`
+}