+++ testdata/panic_ts_tag_uninferrable/input.go
@@ -1,0 +1,6 @@
+package tstaguninferrable
+
+//gents:export
+type Foo struct {
+	When string `json:"when" ts:"Date"`
+}