+++ testdata/panic_ts_tag_with_string_flag/input.go
@@ -1,0 +1,6 @@
+package tstagwithstringflag
+
+//gents:export
+type Foo struct {
+	N int64 `json:"n,string" ts:"string"`
+}